renamed log_verbose module to logger
